Develop merge

+++ Assets/Scripts/01.Animal/Interfaces/IConductable.cs
@@ -3,5 +3,5 @@ using System.Numerics;
 public interface IConductable
 {
     public int Stamina { get; set; }
-    public BigInteger Workload { get; set; }
+    public int Workload { get; set; }
 }
